Fix for Play Framework Less.js integration
I had this file compilation failing using Play Framework due to the missing definition of column-width, gutter-width and total-columns. 

I think it's because Play uses an old version of Less.js (<= 1.3.1, but not really sure if there is any other validity check in the framework itself that fails).
 
I found that adding a default value of 0 to all of these works just fine.